requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Google Ads API
google-ads==23.1.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-multipart==0.0.6

# Logging and Monitoring
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Date and Time
python-dateutil==2.8.2
pytz==2023.3

# Validation
email-validator==2.1.0

